+++ labelbox/schema/project.py
@@ -7,9 +7,9 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, Dict, Iterable, List, Optional, Union
 from urllib.parse import urlparse
 
-from labelbox import parser
 import requests
 
+from labelbox import parser
 from labelbox import utils
 from labelbox.exceptions import (InvalidQueryError, LabelboxError,
                                  ProcessingWaitTimeout, ResourceConflict,
@@ -19,6 +19,7 @@ from labelbox.orm.db_object import DbObject, Deletable, Updateable
 from labelbox.orm.model import Entity, Field, Relationship
 from labelbox.pagination import PaginatedCollection
 from labelbox.schema.consensus_settings import ConsensusSettings
+from labelbox.schema.create_batches_task import CreateBatchesTask
 from labelbox.schema.data_row import DataRow
 from labelbox.schema.export_filters import ProjectExportFilters, validate_datetime, build_filters
 from labelbox.schema.export_params import ProjectExportParams
@@ -26,8 +27,8 @@ from labelbox.schema.media_type import MediaType
 from labelbox.schema.queue_mode import QueueMode
 from labelbox.schema.resource_tag import ResourceTag
 from labelbox.schema.task import Task
-from labelbox.schema.user import User
 from labelbox.schema.task_queue import TaskQueue
+from labelbox.schema.user import User
 
 if TYPE_CHECKING:
     from labelbox import BulkImportRequest
@@ -721,14 +722,19 @@ class Project(DbObject, Updateable, Deletable):
         consensus_settings: Optional[Dict[str, float]] = None,
         global_keys: Optional[List[str]] = None,
     ):
-        """Create a new batch for a project. One of `global_keys` or `data_rows` must be provided but not both.
+        """
+        Creates a new batch for a project. One of `global_keys` or `data_rows` must be provided, but not both. A
+            maximum of 100,000 data rows can be added to a batch.
 
         Args:
             name: a name for the batch, must be unique within a project
             data_rows: Either a list of `DataRows` or Data Row ids. 
             global_keys: global keys for data rows to add to the batch. 
             priority: An optional priority for the Data Rows in the Batch. 1 highest -> 5 lowest
-            consensus_settings: An optional dictionary with consensus settings: {'number_of_labels': 3, 'coverage_percentage': 0.1}
+            consensus_settings: An optional dictionary with consensus settings: {'number_of_labels': 3,
+                'coverage_percentage': 0.1}
+
+        Returns: the created batch
         """
 
         # @TODO: make this automatic?
@@ -772,6 +778,156 @@ class Project(DbObject, Updateable, Deletable):
         else:
             return self._create_batch_sync(name, dr_ids, global_keys, priority,
                                            consensus_settings)
+
+    def create_batches(
+        self,
+        name_prefix: str,
+        data_rows: Optional[List[Union[str, DataRow]]] = None,
+        global_keys: Optional[List[str]] = None,
+        priority: int = 5,
+        consensus_settings: Optional[Dict[str, float]] = None,
+    ) -> CreateBatchesTask:
+        """
+        Creates batches for a project from a list of data rows. One of `global_keys` or `data_rows` must be provided,
+        but not both. When more than 100k data rows are specified and thus multiple batches are needed, the specific
+        batch that each data row will be placed in is undefined.
+
+        Batches will be created with the specified name prefix and a unique suffix. The suffix will be a 4-digit
+        number starting at 0000. For example, if the name prefix is "batch" and 3 batches are created, the names
+        will be "batch0000", "batch0001", and "batch0002". This method will throw an error if a batch with the same
+        name already exists.
+
+        Args:
+            name_prefix: a prefix for the batch names, must be unique within a project
+            data_rows: Either a list of `DataRows` or Data Row ids.
+            global_keys: global keys for data rows to add to the batch.
+            priority: An optional priority for the Data Rows in the Batch. 1 highest -> 5 lowest
+            consensus_settings: An optional dictionary with consensus settings: {'number_of_labels': 3,
+                'coverage_percentage': 0.1}
+
+        Returns: a task for the created batches
+        """
+
+        if self.queue_mode != QueueMode.Batch:
+            raise ValueError("Project must be in batch mode")
+
+        dr_ids = []
+        if data_rows is not None:
+            for dr in data_rows:
+                if isinstance(dr, Entity.DataRow):
+                    dr_ids.append(dr.uid)
+                elif isinstance(dr, str):
+                    dr_ids.append(dr)
+                else:
+                    raise ValueError(
+                        "`data_rows` must be DataRow ids or DataRow objects")
+
+        self._wait_until_data_rows_are_processed(
+            dr_ids, global_keys, self._wait_processing_max_seconds)
+
+        if consensus_settings:
+            consensus_settings = ConsensusSettings(**consensus_settings).dict(
+                by_alias=True)
+
+        method = 'createBatches'
+        mutation_str = """mutation %sPyApi($projectId: ID!, $input: CreateBatchesInput!) {
+                                      project(where: {id: $projectId}) {
+                                        %s(input: $input) {
+                                          tasks {
+                                            batchUuid
+                                            taskId
+                                          }
+                                        }
+                                      }
+                                    }
+                                """ % (method, method)
+
+        params = {
+            "projectId": self.uid,
+            "input": {
+                "batchNamePrefix": name_prefix,
+                "dataRowIds": dr_ids,
+                "globalKeys": global_keys,
+                "priority": priority,
+                "consensusSettings": consensus_settings
+            }
+        }
+
+        tasks = self.client.execute(
+            mutation_str, params, experimental=True)["project"][method]["tasks"]
+        batch_ids = [task["batchUuid"] for task in tasks]
+        task_ids = [task["taskId"] for task in tasks]
+
+        return CreateBatchesTask(self.client, self.uid, batch_ids, task_ids)
+
+    def create_batches_from_dataset(
+        self,
+        name_prefix: str,
+        dataset_id: str,
+        priority: int = 5,
+        consensus_settings: Optional[Dict[str,
+                                          float]] = None) -> CreateBatchesTask:
+        """
+        Creates batches for a project from a dataset, selecting only the data rows that are not already added to the
+        project. When the dataset contains more than 100k data rows and multiple batches are needed, the specific batch
+        that each data row will be placed in is undefined. Note that data rows may not be immediately available for a
+        project after being added to a dataset; use the `_wait_until_data_rows_are_processed` method to ensure that
+        data rows are available before creating batches.
+
+        Batches will be created with the specified name prefix and a unique suffix. The suffix will be a 4-digit
+        number starting at 0000. For example, if the name prefix is "batch" and 3 batches are created, the names
+        will be "batch0000", "batch0001", and "batch0002". This method will throw an error if a batch with the same
+        name already exists.
+
+        Args:
+            name_prefix: a prefix for the batch names, must be unique within a project
+            dataset_id: the id of the dataset to create batches from
+            priority: An optional priority for the Data Rows in the Batch. 1 highest -> 5 lowest
+            consensus_settings: An optional dictionary with consensus settings: {'number_of_labels': 3,
+                'coverage_percentage': 0.1}
+
+        Returns: a task for the created batches
+        """
+
+        if self.queue_mode != QueueMode.Batch:
+            raise ValueError("Project must be in batch mode")
+
+        if consensus_settings:
+            consensus_settings = ConsensusSettings(**consensus_settings).dict(
+                by_alias=True)
+
+        print("Creating batches from dataset %s", dataset_id)
+
+        method = 'createBatchesFromDataset'
+        mutation_str = """mutation %sPyApi($projectId: ID!, $input: CreateBatchesFromDatasetInput!) {
+                                        project(where: {id: $projectId}) {
+                                            %s(input: $input) {
+                                              tasks {
+                                                batchUuid
+                                                taskId
+                                              }
+                                            }
+                                        }
+                                    }
+                                """ % (method, method)
+
+        params = {
+            "projectId": self.uid,
+            "input": {
+                "batchNamePrefix": name_prefix,
+                "datasetId": dataset_id,
+                "priority": priority,
+                "consensusSettings": consensus_settings
+            }
+        }
+
+        tasks = self.client.execute(
+            mutation_str, params, experimental=True)["project"][method]["tasks"]
+
+        batch_ids = [task["batchUuid"] for task in tasks]
+        task_ids = [task["taskId"] for task in tasks]
+
+        return CreateBatchesTask(self.client, self.uid, batch_ids, task_ids)
 
     def _create_batch_sync(self, name, dr_ids, global_keys, priority,
                            consensus_settings):
@@ -843,7 +999,7 @@ class Project(DbObject, Updateable, Deletable):
         add_data_rows_mutation_str = """mutation %sPyApi($projectId: ID!, $input: AddDataRowsToBatchInput!) {
                                       project(where: {id: $projectId}) {
                                         %s(input: $input) {
-                                            taskId
+                                          taskId
                                         }
                                       }
                                     }
@@ -871,29 +1027,7 @@ class Project(DbObject, Updateable, Deletable):
             raise LabelboxError(f"Batch was not created successfully: " +
                                 json.dumps(task.errors))
 
-        # obtain batch entity to return
-        get_batch_str = """query %s($projectId: ID!, $batchId: ID!) {
-                          project(where: {id: $projectId}) {
-                             batches(where: {id: $batchId}) {
-                                nodes {
-                                   %s
-                                }
-                             }
-                        }
-                    }
-                    """ % ("getProjectBatchPyApi",
-                           query.results_query_part(Entity.Batch))
-
-        batch = self.client.execute(
-            get_batch_str, {
-                "projectId": self.uid,
-                "batchId": batch_id
-            },
-            timeout=180.0,
-            experimental=True)["project"]["batches"]["nodes"][0]
-
-        # TODO async endpoints currently do not provide failed_data_row_ids in response
-        return Entity.Batch(self.client, self.uid, batch)
+        return self.client.get_batch(self.uid, batch_id)
 
     def _update_queue_mode(self, mode: "QueueMode") -> "QueueMode":
         """
@@ -1320,12 +1454,35 @@ class Project(DbObject, Updateable, Deletable):
         """ Wait until all the specified data rows are processed"""
         start_time = datetime.now()
 
+        max_data_rows_per_poll = 100_000
+        if data_row_ids is not None:
+            for i in range(0, len(data_row_ids), max_data_rows_per_poll):
+                chunk = data_row_ids[i:i + max_data_rows_per_poll]
+                self._poll_data_row_processing_status(
+                    chunk, [], start_time, wait_processing_max_seconds,
+                    sleep_interval)
+
+        if global_keys is not None:
+            for i in range(0, len(global_keys), max_data_rows_per_poll):
+                chunk = global_keys[i:i + max_data_rows_per_poll]
+                self._poll_data_row_processing_status(
+                    [], chunk, start_time, wait_processing_max_seconds,
+                    sleep_interval)
+
+    def _poll_data_row_processing_status(
+            self,
+            data_row_ids: List[str],
+            global_keys: List[str],
+            start_time: datetime,
+            wait_processing_max_seconds: int = _wait_processing_max_seconds,
+            sleep_interval=30):
+
         while True:
             if (datetime.now() -
                     start_time).total_seconds() >= wait_processing_max_seconds:
                 raise ProcessingWaitTimeout(
-                    "Maximum wait time exceeded while waiting for data rows to be processed. Try creating a batch a bit later"
-                )
+                    """Maximum wait time exceeded while waiting for data rows to be processed. 
+                    Try creating a batch a bit later""")
 
             all_good = self.__check_data_rows_have_been_processed(
                 data_row_ids, global_keys)